layoutを分割
