+++ apps/frontend/src/app/(site)/layout.tsx
@@ -1,0 +1,46 @@
+import type { Metadata } from 'next'
+
+import { BaseLayout } from '~/components/Layout/BaseLayout'
+import { SITE_NAME } from '~/app/layout'
+
+export const metadata: Metadata = {
+  title: {
+    default: SITE_NAME,
+    template: `%s | ${SITE_NAME}`,
+  },
+  description: '',
+  openGraph: {
+    title: SITE_NAME,
+    description: '',
+    url: '',
+    siteName: SITE_NAME,
+    images: [
+      {
+        url: '',
+        width: 1200,
+        height: 630,
+        alt: '',
+      },
+    ],
+    type: 'website',
+  },
+  robots: {
+    index: true,
+    follow: true,
+    nocache: false,
+  },
+  twitter: {
+    card: 'summary_large_image',
+    title: SITE_NAME,
+    description: '',
+    images: [''],
+  },
+}
+
+export default function SiteLayout({
+  children,
+}: {
+  children: React.ReactNode
+}) {
+  return <BaseLayout>{children}</BaseLayout>
+}
