+++ apps/frontend/src/app/(site)/shop/[shopId]/page.tsx
