+++ apps/frontend/src/app/admin/page.tsx
@@ -1,0 +1,5 @@
+const Page = () => {
+  return <div>hoge</div>
+}
+
+export default Page
