+++ apps/frontend/src/components/Admin/Layout/BaseLayout/index.tsx
@@ -1,0 +1,13 @@
+import type { ReactNode } from 'react'
+
+type Props = {
+  children: ReactNode
+}
+
+export const BaseLayout = ({ children }: Props): ReactNode => {
+  return (
+    <>
+      <main>{children}</main>
+    </>
+  )
+}
